separate link components and move into dedicated link component folder
