+++ src/components/link/index.js
@@ -1,0 +1,4 @@
+export * from './link'
+export * from './link-icon'
+export * from './external-link'
+export * from './mail-to-link'
