+++ src/components/link/mail-to-link.js
@@ -1,0 +1,14 @@
+import { Fragment } from 'react'
+import { BaseLinkPropTypes } from './'
+import { MailtoLinkIcon } from './'
+
+export const MailtoLink = ({ to, children }) => {
+  return (
+    <Fragment>
+      <a href={ to }>{ children }</a>
+      <MailtoLinkIcon />
+    </Fragment>
+  )
+}
+
+MailtoLink.propTypes = { ...BaseLinkPropTypes }
